feat(#584): moved all spaces stuff to spaces folder and removed membership folder
