+++ packages/epics/src/index.ts
@@ -3,7 +3,6 @@ export * from './governance';
 export * from './people';
 export * from './proposals';
 export * from './agreements';
-export * from './membership';
 export * from './treasury';
 export * from './discussions';
 export * from './common/constants';
