--- packages/epics/src/membership/components/index.ts
@@ -1,3 +1,0 @@
-export * from './outer-spaces';
-export * from './inner-spaces';
-export * from './subspace-detail';
