--- packages/epics/src/membership/components/inner-spaces/index.ts
@@ -1,3 +1,0 @@
-export * from './inner-space-card';
-export * from './inner-spaces-list';
-export * from './inner-spaces-section';
