--- packages/epics/src/membership/components/outer-spaces/index.ts
@@ -1,3 +1,0 @@
-export * from './outer-space-card';
-export * from './outer-spaces-list';
-export * from './outer-spaces-section';
