--- packages/epics/src/membership/components/subspace-detail/index.ts
@@ -1,1 +1,0 @@
-export * from './subspace-detail';
