--- packages/epics/src/membership/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './use-subspace-by-slug';
-export * from './types';
