+++ packages/epics/src/spaces/components/index.ts
@@ -2,3 +2,10 @@ export * from './space-card';
 export * from './space-group-slider';
 export * from './space-search';
 export * from './create-space-form';
+export * from './inner-space-card';
+export * from './inner-spaces-list';
+export * from './inner-spaces-section';
+export * from './outer-space-card';
+export * from './outer-spaces-list';
+export * from './outer-spaces-section';
+export * from './subspace-detail';
