+++ packages/epics/src/spaces/components/inner-spaces-section.tsx
@@ -2,7 +2,7 @@
 import { FC } from 'react';
 import { InnerSpacesList } from './inner-spaces-list';
 import { Text } from '@radix-ui/themes';
-import { useSpacesSection } from '../../hooks/use-spaces-section';
+import { useSpacesSection } from '../hooks';
 import { SectionFilter, SectionLoadMore } from '@hypha-platform/ui/server';
 import { Button } from '@hypha-platform/ui';
 import { PlusIcon } from '@radix-ui/react-icons';
