+++ packages/epics/src/spaces/components/outer-space-card.tsx
