+++ packages/epics/src/spaces/components/outer-spaces-section.tsx
@@ -2,7 +2,7 @@
 import { FC } from 'react';
 import { OuterSpacesList } from './outer-spaces-list';
 import { Text } from '@radix-ui/themes';
-import { useSpacesSection } from '../../hooks/use-spaces-section';
+import { useSpacesSection } from '../hooks';
 import { SectionFilter, SectionLoadMore } from '@hypha-platform/ui/server';
 import { Button } from '@hypha-platform/ui';
 import { PlusIcon } from '@radix-ui/react-icons';
